+++ src/games/index.ts
@@ -27,6 +27,7 @@ import { FendoGame, IFendoState } from "./fendo";
 import { ArchimedesGame, IArchimedesState } from "./archimedes";
 import { ZolaGame, IZolaState } from "./zola";
 import { MonkeyQueenGame, IMonkeyQueenState } from "./monkey";
+import { DipoleGame, IDipoleState } from "./dipole";
 
 export {
     APGamesInformation, GameBase, IAPGameState,
@@ -56,6 +57,7 @@ export {
     ArchimedesGame, IArchimedesState,
     ZolaGame, IZolaState,
     MonkeyQueenGame, IMonkeyQueenState,
+    DipoleGame, IDipoleState
 };
 
 const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof CannonGame |
@@ -65,12 +67,12 @@ const games = new Map<string, typeof AmazonsGame | typeof BlamGame | typeof Cann
                               typeof PikemenGame | typeof OrdoGame | typeof AttangleGame |
                               typeof AccastaGame | typeof EpamGame | typeof TaijiGame | typeof BreakthroughGame |
                               typeof FabrikGame | typeof ManalathGame | typeof UrbinoGame | typeof FendoGame |
-                              typeof ArchimedesGame | typeof ZolaGame | typeof MonkeyQueenGame>();
+                              typeof ArchimedesGame | typeof ZolaGame | typeof MonkeyQueenGame | typeof DipoleGame>();
 // Manually add each game to the following array
 [AmazonsGame, BlamGame, CannonGame, MchessGame, HomeworldsGame, EntropyGame, VolcanoGame, MvolcanoGame,
     ChaseGame, AbandeGame, CephalopodGame, LinesOfActionGame, PikemenGame, OrdoGame, AttangleGame,
     AccastaGame, EpamGame, TaijiGame, BreakthroughGame, FabrikGame, ManalathGame, UrbinoGame, FendoGame,
-    ArchimedesGame, ZolaGame, MonkeyQueenGame].forEach((g) => {
+    ArchimedesGame, ZolaGame, MonkeyQueenGame, DipoleGame].forEach((g) => {
     if (games.has(g.gameinfo.uid)) {
         throw new Error("Another game with the UID '" + g.gameinfo.uid + "' has already been used. Duplicates are not allowed.");
     }
@@ -133,6 +135,8 @@ export const GameFactory = (game: string, ...args: any[]): GameBase|undefined =>
             return new ZolaGame(...args);
         case "monkey":
             return new MonkeyQueenGame(...args);
+        case "dipole":
+            return new DipoleGame(...args);
     }
     return;
 }